+++ bot/src/main/kotlin/org/epilink/bot/db/LinkBanLogic.kt
@@ -32,5 +32,4 @@ internal class LinkBanLogicImpl : KoinComponent, LinkBanLogic {
         val expiry = ban.expiresOn ?: return true
         return expiry.isAfter(Instant.now())
     }
-
 }