+++ web/src/components/Route.vue
@@ -11,7 +11,7 @@
 <template>
     <li class="route">
         <router-link v-if="r.route" :to="{ name: r.route }" v-html="$t(`layout.navigation.${r.route}`)" />
-        <a v-if="r.url" :href="r.url" target="_blank">{{ r.name }}</a>
+        <a v-if="r.url" :href="r.url" rel="noreferrer" target="_blank">{{ r.name }}</a>
     </li>
 </template>