+++ web/src/views/About.vue
@@ -11,13 +11,13 @@
 <template>
     <div id="about">
         <div id="banner">
-            <img id="logo" src="../../assets/logo.svg"/>
+            <img id="logo" alt="Logo" src="../../assets/logo.svg"/>
             <h1 class="title">EpiLink</h1>
         </div>
 
         <div id="button-container">
-            <a id="sources" href="https://github.com/Litarvan/EpiLink" target="_blank">
-                <img id="github" src="../../assets/github.png" />
+            <a id="sources" href="https://github.com/EpiLink/EpiLink" rel="noreferrer" target="_blank">
+                <img id="github" alt="GitHub" src="../../assets/github.png" />
                 {{ $t('about.sources') }}
             </a>
         </div>
@@ -32,13 +32,13 @@
             <h2 class="title" v-html="$t('about.authors')" />
 
             <div id="authors">
-                <a class="author" href="https://github.com/utybo" target="_blank">
-                    <img class="avatar" src="../../assets/utybo.png" />
+                <a class="author" href="https://github.com/utybo" rel="noreferrer" target="_blank">
+                    <img class="avatar" alt="Zoroark" src="../../assets/utybo.png" />
                     <span class="name">Matthieu 'Zoroark' Stombellini</span>
                     <span class="role">Back-end</span>
                 </a>
-                <a class="author" href="https://github.com/Litarvan" target="_blank">
-                    <img class="avatar" src="../../assets/litarvan.png" />
+                <a class="author" href="https://github.com/Litarvan" rel="noreferrer" target="_blank">
+                    <img class="avatar" alt="Litarvan" src="../../assets/litarvan.png" />
                     <span class="name">Adrien 'Litarvan' Navratil</span>
                     <span class="role">Front-end</span>
                 </a>