+++ web/src/views/Redirect.vue
@@ -69,7 +69,7 @@
 </script>
 
 <style lang="scss" scoped>
-    @import 'src/styles/mixins';
+    @import '../styles/mixins';
     @import '../styles/vars';
 
     #redirect {